# check_interrupt: add a branch hint